requirements.txt.txt
python-telegram-bot==20.7
nest_asyncio
beautifulsoup4
requests
apscheduler
